+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - GeoJSON to JSON and JSON to GeoJSON conversion
+- Improved in GeoJSON to JSON:
+    - Makes Phases into references & populates phases array (Nodes, Spans)
 
 ## [0.2.0] - 2022-11-03